requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.41
Flask-Login==0.5.0
Flask-WTF==0.15.1
Python-Dotenv==0.19.0
Supabase==1.0.3
Gunicorn==20.1.0
Email-Validator==1.1.3
Werkzeug==2.0.1